+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Bundle a couple of print functions compatible with the print middleware (see `nrepl.util.print`).
 * [#174](https://github.com/nrepl/nrepl/issues/174): Provide a built-in `completions` op.
-* [#143](https://github.com/nrepl/nrepl/issues/143): Added a middleware that allows dynamic loading/unloading of middlewares while the server is running.
+* [#143](https://github.com/nrepl/nrepl/issues/143): Added a middleware that allows dynamic loading/unloading of middleware while the server is running.
 * [#180](https://github.com/nrepl/nrepl/issues/180): Provide a built-in `lookup` op.
 
 ### Bugs fixed